Download to folder for reusability and send as url if url is defined

+++ pixivdownloader/bot/command.py
@@ -1,15 +1,16 @@
 from io import BytesIO
 from itertools import chain, islice
 from pathlib import Path
-from tempfile import TemporaryDirectory
+from urllib.parse import urlparse
 import logging
+import re
 
 from pixiv.downloader import PixivDownloader, PixivDownloaderError
 from telegram import Bot, Update, InputMediaPhoto
 from telegram.ext import run_async, MessageHandler, Filters
 
 from pixivdownloader.bot.bot import main_bot
-from pixivdownloader.bot.settings import PIXIV_USERNAME, PIXIV_PASSWORD
+from pixivdownloader.bot.settings import PIXIV_USERNAME, PIXIV_PASSWORD, URL, DOWNLOAD_TO
 
 
 class Command:
@@ -17,6 +18,8 @@ class Command:
         self.logger = logging.getLogger(self.__class__.__name__)
 
         self.client = PixivDownloader(username=PIXIV_USERNAME, password=PIXIV_PASSWORD)
+        self.out_dir = Path(DOWNLOAD_TO)
+        self.out_dir.mkdir(parents=True, exist_ok=True)
 
         main_bot.add_command(name='start', func=self.start)
         main_bot.add_command(MessageHandler, func=self.downloader, filters=Filters.text)
@@ -41,37 +44,50 @@ Just send me a link or the id of the post and I'll give you the images / videos.
         bytes.seek(0)
         return bytes
 
+    def _file_to_upload(self, path: Path) -> str or BytesIO:
+        if URL:
+            return URL.rstrip('/') + '/' + str(path.relative_to(self.out_dir))
+        else:
+            return self._file_to_bytes(path)
+
+
     @run_async
-    def _download(self, bot, update, url):
+    def _download(self, bot, update, id):
         chat_id = update.effective_chat.id
         message = update.effective_message
         message_id = message.message_id
 
-        with TemporaryDirectory() as dir:
+        out_path = self.out_dir / id
+        downloadin_msg = None
+        if out_path.is_dir():
+            downloads = sorted(out_path.glob('*.*'))
+        else:
+            out_path.mkdir()
+
             try:
-                downloads = self.client.download_by_url(url, dir)
-                downloadin_msg = message.reply_text(f'Downloading {url}', reply_to_message_id=message_id,
-                                                          disable_web_page_preview=True)
+                downloads = self.client.download_by_id(id, out_path)
+                downloadin_msg = message.reply_markdown(f'Downloading `{id}`', reply_to_message_id=message_id)
                 downloadin_msg = downloadin_msg.result()
             except PixivDownloaderError:
-                message.reply_text(f'Post ({url}) not found', reply_to_message_id=message_id, disable_web_page_preview=True)
+                message.reply_markdown(f'Post `{id}` not found', reply_to_message_id=message_id)
                 return
 
-            for index, chunk in enumerate(self._chunks(downloads, 10)):
-                self.logger.info(f'Downloading chuck {index} of {url}')
-                works = list(chunk)
+        for index, chunk in enumerate(self._chunks(downloads, 10)):
+            self.logger.info(f'Getting chuck {index} of {id}')
+            works = list(chunk)
 
-                if len(works) == 1:
-                    work = works[0]
-                    if work.suffix == '.mp4':
-                        bot.send_video(chat_id, self._file_to_bytes(work), reply_to_message_id=message_id, caption=url,
-                                       timeout=60)
-                    else:
-                        bot.send_photo(chat_id, self._file_to_bytes(work), reply_to_message_id=message_id, caption=url)
+            if len(works) == 1:
+                work = works[0]
+                if work.suffix == '.mp4':
+                    bot.send_video(chat_id, self._file_to_upload(work), reply_to_message_id=message_id, caption=id,
+                                   timeout=60)
                 else:
-                    media_group = map(InputMediaPhoto, map(self._file_to_bytes, works))
-                    bot.send_media_group(chat_id, media_group, reply_to_message_id=message_id,
-                                         timeout=120, caption=url)
+                    bot.send_photo(chat_id, self._file_to_upload(work), reply_to_message_id=message_id, caption=id)
+            else:
+                media_group = map(InputMediaPhoto, map(self._file_to_upload, works))
+                bot.send_media_group(chat_id, media_group, reply_to_message_id=message_id,
+                                     timeout=120, caption=id)
+        if downloadin_msg:
             downloadin_msg.delete()
 
     def downloader(self, bot: Bot, update: Update):
@@ -82,7 +98,9 @@ Just send me a link or the id of the post and I'll give you the images / videos.
             return
 
         for url in urls.split('\n'):
-            self._download(bot, update, url)
+            ids = re.findall('(\\d+)', urlparse(url).path)
+            for id in ids:
+                self._download(bot, update, id)
 
 
 command = Command()
